feat: use delete edge hook
