+++ apps/journeys-admin/src/components/Editor/Slider/JourneyFlow/libs/useDeleteEdge/index.ts
@@ -1,0 +1,1 @@
+export { useDeleteEdge } from './useDeleteEdge'
